feat: Refactor Context API. Add Zustand
