+++ src/app/App.tsx
@@ -1,0 +1,10 @@
+import { RouterProvider } from 'react-router-dom'
+import { router } from './router'
+
+function App() {
+  return (
+    <RouterProvider router={router} />
+  )
+}
+
+export default App
