+++ src/app/store.ts
@@ -1,0 +1,31 @@
+import { create } from 'zustand'
+import { persist, devtools } from 'zustand/middleware'
+import type { TaskStore } from '@/shared/types/types'
+
+
+export const useStore = create<TaskStore>()(
+    persist(
+        devtools((set) => ({
+            tasks: [],
+            addTask: (task) =>
+                set((state) => ({
+                    tasks: [...state.tasks, task],
+                }), false, 'task/add'),
+
+            deleteTask: (id) =>
+                set((state) => ({
+                    tasks: state.tasks.filter((task) => task.id !== id),
+                }), false, 'task/delete'),
+
+            updateTask: (id, updatedTask) =>
+                set((state) => ({
+                    tasks: state.tasks.map((task) =>
+                        task.id !== id ? task : updatedTask
+                    ),
+                }), false, 'task/update'),
+        })),
+        {
+            name: 'task-storage', // имя ключа в localStorage
+        }
+    )
+)
