--- src/index.css
@@ -1,109 +1,0 @@
-@import "tailwindcss";
-@import "tw-animate-css";
-
-@custom-variant dark (&:is(.dark *));
-
-@theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: var(--font-geist-sans);
-  --font-mono: var(--font-geist-mono);
-  --font-montserrat: var(--font-montserrat);
-  --color-sidebar-ring: var(--sidebar-ring);
-  --color-sidebar-border: var(--sidebar-border);
-  --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
-  --color-sidebar-accent: var(--sidebar-accent);
-  --color-sidebar-primary-foreground: var(--sidebar-primary-foreground);
-  --color-sidebar-primary: var(--sidebar-primary);
-  --color-sidebar-foreground: var(--sidebar-foreground);
-  --color-sidebar: var(--sidebar);
-  --color-chart-5: var(--chart-5);
-  --color-chart-4: var(--chart-4);
-  --color-chart-3: var(--chart-3);
-  --color-chart-2: var(--chart-2);
-  --color-chart-1: var(--chart-1);
-  --color-ring: var(--ring);
-  --color-input: var(--input);
-  --color-border: var(--border);
-  --color-destructive: var(--destructive);
-  --color-accent-foreground: var(--accent-foreground);
-  --color-accent: var(--accent);
-  --color-muted-foreground: var(--muted-foreground);
-  --color-muted: var(--muted);
-  --color-secondary-foreground: var(--secondary-foreground);
-  --color-secondary: var(--secondary);
-  --color-primary-foreground: var(--primary-foreground);
-  --color-primary: var(--primary);
-  --color-popover-foreground: var(--popover-foreground);
-  --color-popover: var(--popover);
-  --color-card-foreground: var(--card-foreground);
-  --color-card: var(--card);
-  --radius-sm: calc(var(--radius) - 4px);
-  --radius-md: calc(var(--radius) - 2px);
-  --radius-lg: var(--radius);
-  --radius-xl: calc(var(--radius) + 4px);
-}
-
-:root {
-  --radius: 0.625rem;
-  --background: oklch(97.0% 0.0111 180);
-    --foreground: oklch(10.0% 0.037 180);
-    --card: oklch(98.0% 0.0074 180);
-    --card-foreground: oklch(10.0% 0.037 180);
-    --popover: oklch(98.0% 0.0074 180);
-    --popover-foreground: oklch(10.0% 0.037 180);
-    --primary: oklch(70.0% 0.2368 150);
-    --primary-foreground: oklch(98.0% 0.0118 150);
-    --secondary: oklch(50.0% 0.333 35);
-    --secondary-foreground: oklch(10.0% 0.0666 35);
-    --muted: oklch(90.0% 0.0222 180);
-    --muted-foreground: oklch(40.0% 0.0888 180);
-    --accent: oklch(80.0% 0.2072 310);
-    --accent-foreground: oklch(10.0% 0.0518 330);
-    --destructive: oklch(60.0% 0.2664 0);
-    --destructive-foreground: oklch(98.0% 0.0133 0);
-    --border: oklch(85.0% 0.0555 180);
-    --input: oklch(85.0% 0.0555 180);
-    --ring: oklch(40.0% 0.2368 150);
-    --chart-1: oklch(40.0% 0.2368 150);
-    --chart-2: oklch(50.0% 0.333 35);
-    --chart-3: oklch(60.0% 0.2072 330);
-    --chart-4: oklch(50.0% 0.333 200);
-    --chart-5: oklch(50.0% 0.333 50);
-}
-
-.dark {
-  --background: oklch(10.0% 0.0518 200);
-  --foreground: oklch(97.0% 0.0111 180);
-  --card: oklch(12.0% 0.0622 200);
-  --card-foreground: oklch(97.0% 0.0111 180);
-  --popover: oklch(12.0% 0.0622 200);
-  --popover-foreground: oklch(97.0% 0.0111 180);
-  --primary: oklch(70.0% 0.296 150);
-  --primary-foreground: oklch(10.0% 0.0592 150);
-  --secondary: oklch(78.88% 0.148 70.87);
-  --secondary-foreground: oklch(10.0% 0.0666 35);
-  --muted: oklch(20.0% 0.074 200);
-  --muted-foreground: oklch(80.0% 0.0444 180);
-  --accent: oklch(70.0% 0.1554 330);
-  --accent-foreground: oklch(10.0% 0.0518 330);
-  --destructive: oklch(70.0% 0.1998 0);
-  --destructive-foreground: oklch(10.0% 0.0666 0);
-  --border: oklch(30.0% 0.1554 200);
-  --input: oklch(30.0% 0.1554 200);
-  --ring: oklch(50.0% 0.296 150);
-  --chart-1: oklch(50.0% 0.296 150);
-  --chart-2: oklch(60.0% 0.2664 35);
-  --chart-3: oklch(70.0% 0.1554 330);
-  --chart-4: oklch(60.0% 0.2664 200);
-  --chart-5: oklch(60.0% 0.2664 50);
-}
-
-@layer base {
-  * {
-    @apply border-border outline-ring/50;
-  }
-  body {
-    @apply bg-background text-foreground;
-  }
-}
