+++ src/pages/error/ErrorPage.tsx
@@ -1,0 +1,18 @@
+import { useRouteError } from 'react-router'
+
+function ErrorPage() {
+    const error: any = useRouteError()
+    console.error(error)
+
+    return (
+        <div id="error-page">
+            <h1>Oops!</h1>
+            <p>Sorry, an unexpected error has occurred.</p>
+            <p>
+                <i>{error.statusText || error.message}</i>
+            </p>
+        </div>
+    )
+}
+
+export default ErrorPage
