+++ src/pages/task-list/TaskListRender.tsx
@@ -1,0 +1,12 @@
+import TaskList from '@/entities/task/ui/TaskList'
+
+/** 
+Отображает список задач в виде адаптивной сетки карточек (Grid или Flexbox).
+Каждая задача отображается через компонент TaskItem
+*/
+
+function TaskListRender() {
+    return <TaskList />
+}
+
+export default TaskListRender
