+++ src/shared/types/types.ts
@@ -1,0 +1,36 @@
+export type TPriority = 'Low' | 'Medium' | 'High'
+export type TCategory = 'Bug' | 'Feature' | 'Documentation' | 'Refactor' | 'Test'
+export type TStatus = 'To Do' | 'In Progress' | 'Done'
+
+export type TContextArray = Array<TContext>
+
+export type TContext = {
+    id: string
+    header: string
+    description: string
+    priority: TPriority
+    category: TCategory
+    status: TStatus
+
+    createdAt?: string
+    updatedAt?: string
+}
+
+export type TaskStore = {
+    tasks: TContextArray
+    addTask: (task: TContext) => void
+    deleteTask: (id: string) => void
+    updateTask: (id: string, updatedTask: TContext) => void
+}
+
+export type TState = TContextArray
+
+export type TAction =
+    | { type: 'ADD_TASK'; payload: TContext }
+    | { type: 'REMOVE_TASK'; payload: string }
+    | { type: 'UPDATE_TASK'; payload: TContext }
+
+export type TContextType = {
+    state: TState
+    dispatch: React.Dispatch<TAction>
+}
